TDFA compiles with *Orbit
A this point the TDFA can be compiled and supports the
EnterOrbit/LeaveOrbit commands.  It also allows this to be shown, but
only with examineDFA . nfaToDFA instead of show.

This entailed creating the RunState RWS monad for updating the state.
At the moment the enter/leave support is just writing strings
describing it instead of performing any book-keeping.